// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-resolved-ayah-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingResolvedWordDto } from '../models/linking-resolved-word-dto';
export interface LinkingResolvedAyahDto {
  ayahId: number;
  ayahNumber: number;
  matchedQuranWordIds: Array<number>;
  pageFrom: number;
  pageTo: number;
  surahNameArabic: string;
  surahNumber: number;
  verseKey: string;
  words: Array<LinkingResolvedWordDto>;
}